+++ cw-relayer/config.toml
@@ -11,7 +11,7 @@ event_type = "ojo.oracle.v1.EventSetFxRate"
 event_timeout = "1000ms"
 
 # max duration between ticks (to trigger a event rpc change)
-max_tick_timeout = "100s"
+max_tick_timeout = "500s"
 query_timeout = "3000ms"
 
 # max query retries to fetch exchange rates or connect to event rpc at startup
