+++ cw-relayer/relayer/messages.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 
 	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
+	"github.com/cosmos/cosmos-sdk/types"
 
 	"github.com/ojo-network/cw-relayer/relayer/client"
 )
@@ -72,7 +73,7 @@ type (
 	}
 )
 
-func restartQuery(contractAddress, Denom string) ([]client.SmartQuery, error) {
+func genRestartQueries(contractAddress, Denom string) ([]client.SmartQuery, error) {
 	rateData, err := json.Marshal(rateMsg{Ref: symbol{Symbol: Denom}})
 	if err != nil {
 		return nil, err
@@ -111,4 +112,62 @@ func restartQuery(contractAddress, Denom string) ([]client.SmartQuery, error) {
 			},
 		},
 	}, nil
+}
+
+func (r *Relayer) genWasmMsg(msgData []byte) *wasmtypes.MsgExecuteContract {
+	return &wasmtypes.MsgExecuteContract{
+		Sender:   r.relayerClient.RelayerAddrString,
+		Contract: r.contractAddress,
+		Msg:      msgData,
+		Funds:    nil,
+	}
+}
+
+func genRateMsgData(forceRelay bool, msgType MsgType, requestID uint64, resolveTime int64, rates types.DecCoins) (msgData []byte, err error) {
+	msg := Msg{
+		SymbolRates: nil,
+		ResolveTime: resolveTime,
+		RequestID:   requestID,
+	}
+
+	if msgType != RelayHistoricalMedian {
+		for _, rate := range rates {
+			msg.SymbolRates = append(msg.SymbolRates, [2]interface{}{rate.Denom, rate.Amount.Mul(RateFactor).TruncateInt().String()})
+		}
+	}
+
+	switch msgType {
+	case RelayRate:
+		if forceRelay {
+			msgData, err = json.Marshal(MsgForceRelay{Relay: msg})
+		} else {
+			msgData, err = json.Marshal(MsgRelay{Relay: msg})
+		}
+
+	case RelayHistoricalMedian:
+		// collect denom's medians
+		medianRates := map[string][]string{}
+		for _, rate := range rates {
+			medianRates[rate.Denom] = append(medianRates[rate.Denom], rate.Amount.Mul(RateFactor).TruncateInt().String())
+		}
+
+		for denom, medians := range medianRates {
+			msg.SymbolRates = append(msg.SymbolRates, [2]interface{}{denom, medians})
+		}
+
+		if forceRelay {
+			msgData, err = json.Marshal(MsgForceRelayHistoricalMedian{Relay: msg})
+		} else {
+			msgData, err = json.Marshal(MsgRelayHistoricalMedian{Relay: msg})
+		}
+
+	case RelayHistoricalDeviation:
+		if forceRelay {
+			msgData, err = json.Marshal(MsgForceRelayHistoricalDeviation{Relay: msg})
+		} else {
+			msgData, err = json.Marshal(MsgRelayHistoricalDeviation{Relay: msg})
+		}
+	}
+
+	return
 }
