+++ cw-relayer/relayer/relayer.go
@@ -8,7 +8,6 @@ import (
 	"sync"
 	"time"
 
-	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
 	"github.com/cosmos/cosmos-sdk/telemetry"
 	"github.com/cosmos/cosmos-sdk/types"
 	oracletypes "github.com/ojo-network/ojo/x/oracle/types"
@@ -146,9 +145,16 @@ func (r *Relayer) Stop() {
 	<-r.closer.Done()
 }
 
+// incrementIndex increases index to switch to different query rpc
+func (r *Relayer) increment() {
+	r.queryRetries += 1
+	r.index = (r.index + 1) % len(r.queryRPCS)
+	r.logger.Info().Int("rpc index", r.index).Msg("switching query rpc")
+}
+
 // restart queries wasmd chain to fetch latest request, median request and deviation request id
 func (r *Relayer) restart(ctx context.Context) error {
-	queryMsgs, err := restartQuery(r.contractAddress, r.config.Denom)
+	queryMsgs, err := genRestartQueries(r.contractAddress, r.config.Denom)
 	if err != nil {
 		return err
 	}
@@ -184,13 +190,6 @@ func (r *Relayer) restart(ctx context.Context) error {
 	}
 
 	return nil
-}
-
-// incrementIndex increases index to switch to different query rpc
-func (r *Relayer) increment() {
-	r.queryRetries += 1
-	r.index = (r.index + 1) % len(r.queryRPCS)
-	r.logger.Info().Int("rpc index", r.index).Msg("switching query rpc")
 }
 
 func (r *Relayer) setDenomPrices(ctx context.Context, postMedian bool) error {
@@ -232,7 +231,7 @@ func (r *Relayer) setDenomPrices(ctx context.Context, postMedian bool) error {
 
 	var mu sync.Mutex
 	g, _ := errgroup.WithContext(ctx)
-	
+
 	g.Go(func() error {
 		deviationsQueryResponse, err := queryClient.MedianDeviations(ctx, &oracletypes.QueryMedianDeviations{})
 		if err != nil {
@@ -364,62 +363,4 @@ func (r *Relayer) tick(ctx context.Context) error {
 	}
 
 	return nil
-}
-
-func (r *Relayer) genWasmMsg(msgData []byte) *wasmtypes.MsgExecuteContract {
-	return &wasmtypes.MsgExecuteContract{
-		Sender:   r.relayerClient.RelayerAddrString,
-		Contract: r.contractAddress,
-		Msg:      msgData,
-		Funds:    nil,
-	}
-}
-
-func genRateMsgData(forceRelay bool, msgType MsgType, requestID uint64, resolveTime int64, rates types.DecCoins) (msgData []byte, err error) {
-	msg := Msg{
-		SymbolRates: nil,
-		ResolveTime: resolveTime,
-		RequestID:   requestID,
-	}
-
-	if msgType != RelayHistoricalMedian {
-		for _, rate := range rates {
-			msg.SymbolRates = append(msg.SymbolRates, [2]interface{}{rate.Denom, rate.Amount.Mul(RateFactor).TruncateInt().String()})
-		}
-	}
-
-	switch msgType {
-	case RelayRate:
-		if forceRelay {
-			msgData, err = json.Marshal(MsgForceRelay{Relay: msg})
-		} else {
-			msgData, err = json.Marshal(MsgRelay{Relay: msg})
-		}
-
-	case RelayHistoricalMedian:
-		// collect denom's medians
-		medianRates := map[string][]string{}
-		for _, rate := range rates {
-			medianRates[rate.Denom] = append(medianRates[rate.Denom], rate.Amount.Mul(RateFactor).TruncateInt().String())
-		}
-
-		for denom, medians := range medianRates {
-			msg.SymbolRates = append(msg.SymbolRates, [2]interface{}{denom, medians})
-		}
-
-		if forceRelay {
-			msgData, err = json.Marshal(MsgForceRelayHistoricalMedian{Relay: msg})
-		} else {
-			msgData, err = json.Marshal(MsgRelayHistoricalMedian{Relay: msg})
-		}
-
-	case RelayHistoricalDeviation:
-		if forceRelay {
-			msgData, err = json.Marshal(MsgForceRelayHistoricalDeviation{Relay: msg})
-		} else {
-			msgData, err = json.Marshal(MsgRelayHistoricalDeviation{Relay: msg})
-		}
-	}
-
-	return
 }
